Replace getid3 with stripped down version
getID3 is the library that we use in core to read the ID3 tags. It works great for that, however it also includes a ton of magic that also includes invoking external scripts etc.

This imports "lukasreschke/id3parser" which is basically getID3 with all the features removed that we don't need but would potentially make ownCloud insecure.

Also makes me a lot happier since I don't have to argue every time with customers why the library calls so many external sinkholes :see_no_evil:

Or to say it in other words:

> 105 changed files with 85 additions and 21,967 deletions.