chore: set version in composer.json
When running dump-autoload, then composer tries to find a matching version number.

Sometimes the VCS detection does not work and the generated version number changes from the commit hash to the fallback and back to the generated version when someone else updates the package.

We are now using a version number in composer.json to avoid a version number ping-pong.

Signed-off-by: Daniel Kesselberg <mail@danielkesselberg.de>