chore: remove explicit dependency on `symfony/polyfill-intl-grapheme`
We do not use the grapheme extension, it was added together with IDN
polyfills, but we only use the IDN ones. So we do not need an explicit dependency here.

It will still be included for the moment, as we have some dependencies
that require it, so it is now an implicit dependency.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>